+++ benchmark/scripts/runners/run_multi_process_benchmark.sh
@@ -33,12 +33,31 @@ if [ $# -lt 1 ]; then
   exit 1
 fi
 
+# Set default timeout for how long the script should wait after spawning the router
+# before running the benchmarks.
+if [[ -z "$ZENOH_ROUTER_WAIT_TIMEOUT" ]]; then
+  ZENOH_ROUTER_WAIT_TIMEOUT=1.0
+fi
+
+
 CONFIG_FILE=$1
 shift # Shift arguments to parse options
 
 if [ ! -f "$CONFIG_FILE" ]; then
   echo -e "\033[31m[ERROR] Configuration file '$CONFIG_FILE' not found!\033[0m"
   exit 1
+fi
+
+# Ensure scripts directory is set if not provided externally.
+if [ -z "${ROS2_BENCHMARK_SCRIPTS_DIR}" ]; then
+  THIS_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
+  ROS2_BENCHMARK_SCRIPTS_DIR="${THIS_DIR}/.."
+fi
+
+# Set default output path if not provided externally.
+if [[ -z "$ROS2_BENCHMARK_OUTPUT_DIR" ]]; then
+  current_date=$(date +"%d_%m_%y_%Hh%M")
+  ROS2_BENCHMARK_OUTPUT_DIR="/benchmark_results/results_${current_date}"
 fi
 
 # Set default test duration if not provided externally.
@@ -106,6 +125,8 @@ if [[ -z "${RMW_LIST}" || -z "${TOPOLOGY1}" ]]; then
   exit 1
 fi
 
+RUNNER_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
+
 # --- Output Directory Setup ---
 OUTPUT_DIR="${ROS2_BENCHMARK_OUTPUT_DIR}/${OUTPUT_DIR_NAME}"
 echo "Results will be stored in: $OUTPUT_DIR"
@@ -136,6 +157,19 @@ for RMW in "${RMW_LIST[@]}"; do
 
     # Loop through the paired topologies defined in the config file.
     for i in "${!TOPOLOGY1[@]}"; do
+      if [[ "$RMW" == "zenoh" ]]; then
+        # Automatically start the router in the background
+        echo "Detected that $RMW is being benchmarked. Spawning router..."
+
+        ${RUNNER_DIR}/run_zenoh_router.sh ${ZENOH_ROUTER_CONFIG_URI} &
+
+        # Wait for the router to come online
+        sleep ${ZENOH_ROUTER_WAIT_TIMEOUT}
+
+        ROUTER_PID=$(pgrep zenohd)
+        echo "Spawned zenoh router with PID ${ROUTER_PID}"
+      fi
+
       T1="${TOPOLOGY1[i]}"
       T2="${TOPOLOGY2[i]}"
       RES="${RESULTS[i]}"
@@ -176,6 +210,18 @@ for RMW in "${RMW_LIST[@]}"; do
 
       eval "$COMMAND"
 
+      if [[ -n ${ROUTER_PID} ]]; then 
+        echo "Stopping zenoh router with PID $ROUTER_PID"
+        kill ${ROUTER_PID}
+        while kill -0 "${ROUTER_PID}">/dev/null 2>&1; do
+            echo "Waiting for zenoh router to exit..."
+            sleep 0.1
+        done        
+        echo "Stopped zenoh router with PID $ROUTER_PID"
+        unset $ROUTER_PID
+      fi
+
+
       if [ $? -ne 0 ]; then
         echo -e "\033[31m[ERROR] Command failed: $COMMAND\033[0m"
         exit 1
@@ -185,6 +231,12 @@ for RMW in "${RMW_LIST[@]}"; do
       echo "     Moving log files to $RESULT_FOLDER"
       mv ./*log "$RESULT_FOLDER"
     done
+        # Unset environment variables at the end of the loop to avoid side effects.
+    unset FASTRTPS_DEFAULT_PROFILES_FILE
+    unset RMW_FASTRTPS_USE_QOS_FROM_XML
+    unset CYCLONEDDS_URI
+    unset ZENOH_ROUTER_CONFIG_URI
+    unset ZENOH_SESSION_CONFIG_URI
   done
 done