+++ benchmark/scripts/runners/run_single_process_benchmark.sh
@@ -31,12 +31,31 @@ if [ $# -lt 1 ]; then
   exit 1
 fi
 
+# Set default timeout for how long the script should wait after spawning the router
+# before running the benchmarks.
+if [[ -z "$ZENOH_ROUTER_WAIT_TIMEOUT" ]]; then
+  ZENOH_ROUTER_WAIT_TIMEOUT=1.0
+fi
+
+
 CONFIG_FILE=$1
 shift # Shift arguments to parse options
 
 if [ ! -f "$CONFIG_FILE" ]; then
   echo -e "\033[31m[ERROR] Configuration file '$CONFIG_FILE' not found!\033[0m"
   exit 1
+fi
+
+# Ensure scripts directory is set if not provided externally.
+if [ -z "${ROS2_BENCHMARK_SCRIPTS_DIR}" ]; then
+  THIS_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
+  ROS2_BENCHMARK_SCRIPTS_DIR="${THIS_DIR}/.."
+fi
+
+# Set default output path if not provided externally.
+if [[ -z "$ROS2_BENCHMARK_OUTPUT_DIR" ]]; then
+  current_date=$(date +"%d_%m_%y_%Hh%M")
+  ROS2_BENCHMARK_OUTPUT_DIR="/benchmark_results/results_${current_date}"
 fi
 
 # Set default test duration if not provided externally.
@@ -104,6 +123,8 @@ if [[ -z "${RMW_LIST}" || -z "${TOPOLOGIES}" ]]; then
   exit 1
 fi
 
+RUNNER_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
+
 # --- Output Directory Setup ---
 # Append a suffix to the output directory if running in remote host mode.
 OUTPUT_DIR="${ROS2_BENCHMARK_OUTPUT_DIR}/${OUTPUT_DIR_NAME}"
@@ -144,6 +165,19 @@ for RMW in "${RMW_LIST[@]}"; do
 
     # Loop through each topology defined in the config file.
     for TOPOLOGY in "${TOPOLOGIES[@]}"; do
+      if [[ "$RMW" == "zenoh" ]]; then
+        # Automatically start the router in the background
+        echo "Detected that $RMW is being benchmarked. Spawning router..."
+
+        ${RUNNER_DIR}/run_zenoh_router.sh ${ZENOH_ROUTER_CONFIG_URI} &
+
+        # Wait for the router to come online
+        sleep ${ZENOH_ROUTER_WAIT_TIMEOUT}
+
+        ROUTER_PID=$(pgrep zenohd)
+        echo "Spawned zenoh router with PID ${ROUTER_PID}"
+      fi
+
       TEST_CASE_DIR="${OUTPUT_DIR}/${TOPOLOGY}"
       mkdir -p "${TEST_CASE_DIR}"
 
@@ -170,7 +204,6 @@ for RMW in "${RMW_LIST[@]}"; do
             exit 1
         fi
         echo "       Running remote process in '$REMOTE_HOST_MODE' mode..."
-        RUNNER_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
         "${RUNNER_DIR}/run_remote_process.sh" "$RMW_IMPLEMENTATION" "$SCRIPT_FOR_REMOTE_HOST"
       fi
 
@@ -180,6 +213,17 @@ for RMW in "${RMW_LIST[@]}"; do
       echo -e "     Command: \n       $COMMAND"
 
       eval "$COMMAND"
+
+      if [[ -n ${ROUTER_PID} ]]; then 
+        echo "Stopping zenoh router with PID $ROUTER_PID"
+        kill ${ROUTER_PID}
+        while kill -0 "${ROUTER_PID}">/dev/null 2>&1; do
+            echo "Waiting for zenoh router to exit..."
+            sleep 0.1
+        done        
+        echo "Stopped zenoh router with PID $ROUTER_PID"
+        unset $ROUTER_PID
+      fi
 
       if [ $? -ne 0 ]; then
         echo -e "\033[31m[ERROR] Command failed: $COMMAND\033[0m"
@@ -191,6 +235,8 @@ for RMW in "${RMW_LIST[@]}"; do
     unset FASTRTPS_DEFAULT_PROFILES_FILE
     unset RMW_FASTRTPS_USE_QOS_FROM_XML
     unset CYCLONEDDS_URI
+    unset ZENOH_ROUTER_CONFIG_URI
+    unset ZENOH_SESSION_CONFIG_URI
   done
 done